chore(get-item-mixin): Create mixin for generic item fetching
